+++ package.json
@@ -5,10 +5,8 @@
   "main": "src/index.js",
   "leadMaintainer": "Friedel Ziegelmayer <dignifiedquire@gmail.com>",
   "browser": {
-    "./src/hmac/index.js": "./src/hmac/index-browser.js",
-    "./src/keys/ecdh.js": "./src/keys/ecdh-browser.js",
-    "./src/aes/ciphers.js": "./src/aes/ciphers-browser.js",
-    "./src/keys/rsa.js": "./src/keys/rsa-browser.js"
+    "crypto": "crypto-browserify",
+    "./src/keys/keypair.js": "./src/keys/keypair-browser.js"
   },
   "files": [
     "src",
@@ -35,12 +33,10 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "asmcrypto.js": "^2.3.2",
     "asn1.js": "^5.0.1",
     "async": "^2.6.2",
-    "bn.js": "^4.11.8",
-    "browserify-aes": "^1.2.0",
     "bs58": "^4.0.1",
+    "crypto-browserify": "^3.12.0",
     "iso-random-stream": "^1.1.0",
     "keypair": "^1.0.1",
     "libp2p-crypto-secp256k1": "~0.3.0",
@@ -48,12 +44,12 @@
     "node-forge": "~0.7.6",
     "pem-jwk": "^2.0.0",
     "protons": "^1.0.1",
-    "rsa-pem-to-jwk": "^1.1.3",
     "tweetnacl": "^1.0.1",
     "ursa-optional": "~0.9.10"
   },
   "devDependencies": {
     "aegir": "^18.2.2",
+    "bn.js": "^4.11.8",
     "benchmark": "^2.1.4",
     "bundlesize": "~0.17.1",
     "chai": "^4.2.0",
