--- src/aes/ciphers-browser.js
@@ -1,8 +1,0 @@
-'use strict'
-
-const crypto = require('browserify-aes')
-
-module.exports = {
-  createCipheriv: crypto.createCipheriv,
-  createDecipheriv: crypto.createDecipheriv
-}
