+++ src/keys/ecdh.js
@@ -1,41 +1,5 @@
 'use strict'
 
-const crypto = require('crypto')
-const nextTick = require('async/nextTick')
+const webcrypto = require('../webcrypto')
 
-const curves = {
-  'P-256': 'prime256v1',
-  'P-384': 'secp384r1',
-  'P-521': 'secp521r1'
-}
-
-exports.generateEphmeralKeyPair = function (curve, callback) {
-  if (!curves[curve]) {
-    return callback(new Error(`Unkown curve: ${curve}`))
-  }
-  const ecdh = crypto.createECDH(curves[curve])
-  ecdh.generateKeys()
-
-  nextTick(() => callback(null, {
-    key: ecdh.getPublicKey(),
-    genSharedKey (theirPub, forcePrivate, cb) {
-      if (typeof forcePrivate === 'function') {
-        cb = forcePrivate
-        forcePrivate = null
-      }
-
-      if (forcePrivate) {
-        ecdh.setPrivateKey(forcePrivate.private)
-      }
-
-      let secret
-      try {
-        secret = ecdh.computeSecret(theirPub)
-      } catch (err) {
-        return cb(err)
-      }
-
-      nextTick(() => cb(null, secret))
-    }
-  }))
-}
+module.exports = webcrypto ? require('./ecdh-webcrypto') : require('./ecdh-crypto')
