+++ src/keys/keypair-browser.js
@@ -1,0 +1,1 @@
+module.exports = require('keypair')
